- Modificada respuesta de pedidos
- Se organiza la consulta, creacón y actualización de créditos

Related work items: #2337

+++ src/sales/services/orders.service.ts
@@ -35,6 +35,8 @@ import { DocumentTypeStockHistory } from 'src/inventories/dtos/create-stockHisto
 import { StatusProduct } from 'src/products/entities/product.entity';
 import { ActionProductsOrder } from '../dtos/add-products-order-input';
 import { TypePayment } from 'src/treasury/entities/payment.entity';
+import { CreditHistoryService } from 'src/credits/services/credit-history.service';
+import { CreditsService } from 'src/credits/services/credits.service';
 
 const populate = [
 	{
@@ -56,6 +58,8 @@ export class OrdersService {
 		private readonly discountRulesService: DiscountRulersService,
 		private readonly conveyorsService: ConveyorsService,
 		private readonly pointOfSalesService: PointOfSalesService,
+		private readonly creditHistoryService: CreditHistoryService,
+		private readonly creditsService: CreditsService,
 	) {}
 
 	async findAll(
@@ -161,7 +165,8 @@ export class OrdersService {
 			if (lastOrder) {
 				number = lastOrder.number + 1;
 			}
-			return this.orderModel.create({
+
+			const newOrder = await this.orderModel.create({
 				customer,
 				shop,
 				number,
@@ -169,6 +174,11 @@ export class OrdersService {
 				user,
 				pointOfSale: user.pointOfSale._id,
 			});
+
+			return {
+				order: newOrder,
+				credit: null,
+			};
 		} else {
 			if (!user.customer) {
 				throw new BadRequestException('El usuario no pertenece a un cliente');
@@ -196,7 +206,7 @@ export class OrdersService {
 				? user?.customer['addresses']?.find((address) => address?.isMain)
 				: undefined;
 
-		return this.orderModel.create({
+		const newOrder = await this.orderModel.create({
 			customer: user.customer,
 			address,
 			shop,
@@ -206,6 +216,15 @@ export class OrdersService {
 			status: StatusOrder[status],
 			company: new Types.ObjectId(companyId),
 		});
+
+		const credit = await this.creditsService.findOne({
+			customerId: user.customer?._id.toString(),
+		});
+
+		return {
+			order: newOrder,
+			credit,
+		};
 	}
 
 	async update(
@@ -215,6 +234,9 @@ export class OrdersService {
 		companyId: string,
 	) {
 		const order = await this.orderModel.findById(orderId).lean();
+		let credit = await this.creditsService.findOne({
+			customerId: order?.customer.toString(),
+		});
 
 		if (!order) {
 			throw new BadRequestException(
@@ -354,8 +376,14 @@ export class OrdersService {
 						});
 					} else {
 						payments.push(order?.payments[i]);
-
-						//TODO: agregar todo lo correspondiente al crédito
+						const creditHistory =
+							await this.creditHistoryService.addCreditHistory(
+								order?._id?.toString(),
+								total,
+								user,
+								companyId,
+							);
+						credit = creditHistory?.credit;
 					}
 				}
 			}
@@ -400,20 +428,21 @@ export class OrdersService {
 				productId: detail?.product?._id.toString(),
 				quantity: detail?.quantity,
 			}));
-
-			await this.stockHistoryService.deleteStock(
-				{
-					details,
-					documentId: orderId,
-					documentType: DocumentTypeStockHistory.ORDER,
-					warehouseId: order.shop.defaultWarehouse['_id'].toString(),
-				},
-				user,
-				companyId,
-			);
+			if (details) {
+				await this.stockHistoryService.deleteStock(
+					{
+						details,
+						documentId: orderId,
+						documentType: DocumentTypeStockHistory.ORDER,
+						warehouseId: order.shop.defaultWarehouse['_id'].toString(),
+					},
+					user,
+					companyId,
+				);
+			}
 		}
 
-		return this.orderModel.findByIdAndUpdate(
+		const newOrder = await this.orderModel.findByIdAndUpdate(
 			orderId,
 			{
 				$set: { ...dataUpdate, user, conveyor },
@@ -424,6 +453,11 @@ export class OrdersService {
 				lean: true,
 			},
 		);
+
+		return {
+			order: newOrder,
+			credit,
+		};
 	}
 
 	/**
@@ -779,7 +813,7 @@ export class OrdersService {
 			tax,
 		};
 
-		return this.orderModel.findByIdAndUpdate(
+		const newOrder = await this.orderModel.findByIdAndUpdate(
 			orderId,
 			{
 				$set: {
@@ -794,6 +828,15 @@ export class OrdersService {
 				lean: true,
 			},
 		);
+
+		const credit = await this.creditsService.findOne({
+			customerId: user.customer?._id.toString(),
+		});
+
+		return {
+			order: newOrder,
+			credit,
+		};
 	}
 
 	async addPayments({ orderId, payments }: AddPaymentsOrderInput, user: User) {
@@ -949,7 +992,7 @@ export class OrdersService {
 			change,
 		};
 
-		return this.orderModel.findByIdAndUpdate(
+		const newOrder = await this.orderModel.findByIdAndUpdate(
 			orderId,
 			{
 				$set: {
@@ -964,5 +1007,14 @@ export class OrdersService {
 				new: true,
 			},
 		);
+
+		const credit = await this.creditsService.findOne({
+			customerId: user.customer?._id.toString(),
+		});
+
+		return {
+			order: newOrder,
+			credit,
+		};
 	}
 }
